refactor(s3): split objects.rs by op family (2886 -> 57 mod + 5 submodules)
Audit 2026-05-19 P9 continuation. S3 object handlers split:

- write.rs    1145  (put_object / copy_object / restore_object)
- read.rs      629  (get_object / head_object / get_object_attributes)
- list.rs      599  (list_objects_v1 / list_objects_v2 / list_object_versions)
- delete.rs    452  (delete_object / delete_objects)
- website.rs    53  (serve_website_object / serve_website_error)
- mod.rs        57  (mod decls)

Methods are pub(crate) so the service/mod.rs dispatcher reaches them
across the new module boundary. Inner cross-fn helpers reach via
crate::service::{persistence_error, io_to_aws, compute_checksum_streaming,
notifications}.

323 unit tests pass.
